update too many shit
